Change lymph.rpc behaviour and introduce a new decorator raw_rpc
The current lymph.rpc decorator will be renamed lymph.raw_rpc, while
the lymph.rpc will be sepecific to simple and most common use cases
where user can use return to send a RPC message and raises to raise
RemoteError in client side or Nack depending on the raises argument

Docs was changed to reflect the new changes.

Warnings this changes are **not** backward compatible.