Dump threads stack when SIGQUIT is send to process
SIGQUIT signal is meant to be used to produce a core dump when it terminates
the process, JAVA virtual machines already support this behaviour, this change
add it to lymph. This is very usefull feature when it come to test why a process
hang, and specially in a production environment, where when this happen ops can
just restart a service by first killing it using ``kill -3 <pid>`` which produce
a snapshot of all greenlets and os-threads stack printed stderr then as a post mortem
developer can figure out why this happen by analyzing the snapshot.