Retryable event handling
Add a retry option to ``lymph.event(...)`` that specify max number of
reties to do in case an exception is raised when handling the given event.
The way retry work is by republishing the event to the handler queue while setting
a header that specify the number of reties left, this way of requeuing an event
at the opposite to ``message.reject(requeue=True)`` enable us to set a limit of the
numbers of retries.

Note: This feature should only be used when the event handler is guarantee to be idempotent.