Add ability to change runtime log level
It's really helpful to be able to change log level without
downtime to inspect behaviour of running service when tracking
bugs specially. The log level change will take effect for only
a given period of time after that it will be reseted.

Example:
$ lymph change-loglevel demo -n lymph -l DEBUG
Change loglevel of all instances of demo service to DEBUG of lymph logger.